feat(TrainMode): Added the MostRelavantFieldTypes class
This class will be responsible for calculating the BitSets that represent Most Accomodating

FieldType for every fieldName. With every addition of a new FieldType to be accomodated in a

fieldName, the OR operation sets the position for the fieldType being added. In the end (or even in

the intermediate), the Final BitSets can be looked at, and the most accomodating FieldType can be

drawn out from there.